requirements.txt
langchain
OpenAI
python-dotenv
ritetag
huggingface_hub
     
google-generativeai
av
numpy
torch
transformers
pillow
streamlit
pybase64
pathlib
bs4